fix: SSE 코드 일부 수정

+++ sorimaeul/src/main/java/com/usagi/sorimaeul/api/service/SseService.java
@@ -1,8 +1,10 @@
 package com.usagi.sorimaeul.api.service;
 
 import com.usagi.sorimaeul.repository.EmitterRepository;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
@@ -10,33 +12,41 @@ import java.io.*;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class SseService {
+
+	@Value("${sse.time-out}")
+	private final long TIME_OUT;
 
 	private final EmitterRepository emitterRepository;
 
-	// subscibe 연결 설정
-	public SseEmitter subscribe(long userCode, SseEmitter sseEmitter) {
+	public SseEmitter subscribe(long userCode) {
+		SseEmitter sseEmitter = new SseEmitter(TIME_OUT);
 		sseEmitter = emitterRepository.save(userCode, sseEmitter);
 
-		sseEmitter.onCompletion(() -> {
-			emitterRepository.deleteByUserCode(userCode);
-		});
-
+		sseEmitter.onCompletion(() -> emitterRepository.deleteByUserCode(userCode));
 		sseEmitter.onTimeout(sseEmitter::complete);
+		sseEmitter.onError(sseEmitter::completeWithError);
 
-		sendConnect(userCode, sseEmitter);
+		sendToClient(userCode, String.valueOf(userCode), "SSE connect OK");
 
 		return sseEmitter;
 	}
 
-	// sendMethod
-	private void sendConnect(long userCode, SseEmitter sseEmitter) {
-		try {
-			sseEmitter.send(SseEmitter.event()
-					.name("CONNECT")
-					.data("connect OK"));
-		} catch (IOException e) {
-			emitterRepository.deleteByUserCode(userCode);
+	public void sendToClient(long userCode, String emitterKey, Object data) {
+		SseEmitter sseEmitter = emitterRepository.findByUserCode(userCode);
+
+		if (sseEmitter != null) {
+			try {
+				log.info("send to client {}:[{}]", emitterKey, data);
+				sseEmitter.send(SseEmitter.event()
+						.id(String.valueOf(userCode))
+						.name(emitterKey)
+						.data(data, MediaType.APPLICATION_JSON));
+			} catch (IOException e) {
+				log.error("failure to send event, id={}, message={}", userCode, e.getMessage());
+				emitterRepository.deleteByUserCode(userCode);
+			}
 		}
 	}
